Pass in relevant info instead of globally available singleton
